Created setup command and fixing some typo

+++ server/jobs/db_init.ts
@@ -1,14 +1,52 @@
 const sh = require('shelljs')
+import inquirer from 'inquirer'
+import { createSpinner } from 'nanospinner'
+import colors from 'colors'
 
 
-console.log("Starting DB init")
-sh.exec('npx prisma migrate reset --force', { silent: true }).output
-console.log('DB is fully reseted and seeded')
-console.log('Started pushing DB')
+const question = async (): Promise<Boolean> => {
+  const foo = await inquirer.prompt({
+    name: 'accepted',
+    type: 'confirm',
+    message: `Ez a művelet teljesen törli az adatbázist, illetve tartalmát. Biztosan szeretnéd folytatni?`,
+    prefix: '❗️'
+  })
+  return foo.accepted
+}
 
-sh.exec('npx prisma db push --accept-data-loss', { silent: true }).output
+const init = async () => {
+  try {
+    const accepted = await question()
 
-console.log('DB is pushed')
+    if (accepted) {
+      let spinner = createSpinner('Creating database...').start()
+      try {
+        sh.exec('npx prisma migrate reset --force', { silent: true }).output
+        spinner.success()
+      } catch (error: any) {
+        spinner.error()
+        throw new Error(error);
+      }
 
-//console.log('Starting server....')
-//sh.exec('npm run start').output
+      spinner = createSpinner('Seed the database with demo records 🌱').start()
+      try {
+        sh.exec('npx prisma db push --accept-data-loss', { silent: true }).output
+        sh.exec('npm run db_demo_seed', { silent: true }).output
+        spinner.success()
+      } catch (error: any) {
+        spinner.error()
+        throw new Error(error);
+      }
+      console.log(`Done! Run \'${colors.green('npm run start')}\' to start the server.`)
+
+      process.exit(0)
+    } else {
+      process.exit(1)
+    }
+
+  } catch (error) {
+    console.error(error)
+  }
+}
+
+init()
